ai/puzzle/requirements.txt
promptflow
promptflow-tools
azure-cosmos
uuid